pyproject.toml
[project]
name = "schema-data-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "redpheonixx"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]

[tool.poetry]
packages = [{include = "schema_data_generator", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
